Exception service added, get mapping fixed

+++ src/main/java/ru/geekbrains/micecreator/service/prototypes/SimpleTypeService.java
@@ -3,6 +3,9 @@ package ru.geekbrains.micecreator.service.prototypes;
 import ru.geekbrains.micecreator.dto.basic.SearchParams;
 import ru.geekbrains.micecreator.dto.basic.list.ListItemDto;
 import ru.geekbrains.micecreator.dto.basic.prototype.BasicDto;
+import ru.geekbrains.micecreator.exceptions.BadInputException;
+import ru.geekbrains.micecreator.exceptions.DataNotFoundException;
+import ru.geekbrains.micecreator.utils.StringUtils;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -40,7 +43,7 @@ public abstract class SimpleTypeService<D extends BasicDto, E> {
 
 	public D createEntity(D dto) {
 		if (dto == null || dto.getId() != null) {
-			throw new RuntimeException("invalid input!");
+			throw new BadInputException("Invalid input for entity creation: Dto must be not null and have no id.");
 		}
 		dto.setName(nameToStandard(dto.getName()));
 		return mapToDto(save(mapToEntity(dto)));
@@ -48,7 +51,10 @@ public abstract class SimpleTypeService<D extends BasicDto, E> {
 
 	public D editEntity(D dto) {
 		if (dto == null || dto.getId() == null) {
-			throw new RuntimeException("invalid input!");
+			throw new BadInputException("Invalid input for entity change: Dto must be not null and dto.id must be not null.");
+		}
+		if (findById(dto.getId()) == null) {
+			throw new DataNotFoundException(String.format("Can't make changes: no entity with id %s found.", dto.getId()));
 		}
 		dto.setName(nameToStandard(dto.getName()));
 		return mapToDto(save(mapToEntity(dto)));
@@ -56,13 +62,16 @@ public abstract class SimpleTypeService<D extends BasicDto, E> {
 
 	public boolean deleteEntity(Integer id) {
 		if (id == null) {
-			throw new RuntimeException("invalid input!");
+			throw new BadInputException("Invalid input for entity delete operation: id must be not null.");
+		}
+		if (findById(id) == null) {
+			throw new DataNotFoundException(String.format("Can't delete entity: no entity with id %s found.", id));
 		}
 		return deleteById(id);
 	}
 
 	public List<ListItemDto> findBySearchParams(SearchParams params) {
-		if (!params.getNamePart().isBlank()) {
+		if (!StringUtils.isBlank(params.getNamePart())) {
 			return findListDtoByNamePart(params.getNamePart());
 		}
 		return findAllListDto();
@@ -70,15 +79,18 @@ public abstract class SimpleTypeService<D extends BasicDto, E> {
 
 
 	protected String nameToStandard(String name) {
-		if (name.isBlank()) {
-			throw new RuntimeException("invalid input!");
+		if (StringUtils.isBlank(name)) {
+			throw new BadInputException("Invalid input: name must be not blank.");
 		}
 		return name.toUpperCase();
 	}
 
 	protected void checkInputId(Integer id) {
 		if (id == null) {
-			throw new RuntimeException("Invalid Input!");
+			throw new BadInputException("Invalid input for operation: id must be not null.");
+		}
+		if (findById(id) == null) {
+			throw new DataNotFoundException(String.format("Can't execute operation: no entity with id %s found.", id));
 		}
 	}
